ref(api): Separate `Deploy` struct into separate module
