+++ src/api/data_types/mod.rs
@@ -1,3 +1,7 @@
+//! Data types used in the api module
+
 mod chunking;
+mod deploy;
 
 pub use self::chunking::*;
+pub use self::deploy::*;
